+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes
 * [UI]: Changed the user icon in recent events to a basic icon instead of a gravatar (0.12.1)
 * [API]: Fixed a bug with the sequencingObject/sequence file REST endpoint that was throwing 404s when it should read a file. (0.12.2)
 * [Developer]: Added ability to build IRIDA package `.zip` file and `tools-list.yml` for distribution and easy installation of tools in Galaxy.
+* [UI]: Major update to the Project > Samples page.
 
 0.11.0 to 0.12.0
 ----------------